QCMPLUS-35 |Add authentication features and model renaming

+++ src/main/java/com/pmn/qcmplus/exception/RoleNotFoundException.java
@@ -1,7 +1,7 @@
 package com.pmn.qcmplus.exception;
 
 public class RoleNotFoundException extends RuntimeException {
-    public RoleNotFoundException(Long id) {
+    public RoleNotFoundException(Integer id) {
         super("Role with id " + id + " not found.");
     }
 }
